+++ src/app/profile/actions/verify-two-factor-action.ts
@@ -5,9 +5,9 @@ import { z } from 'zod';
 import { eq } from 'drizzle-orm';
 import { db } from '@/db';
 import { users } from '@/db/schema/users';
-import { decryptSecret } from '@/lib/encrypt';
+import { decryptSecret } from '@/lib/encryption';
 import { authActionClient, ActionError } from '@/lib/safe-action';
-import { verifySecret, generateRecoveryCodes } from '@/lib/two-factor';
+import { verifySecret, generateRecoveryCodes } from '@/lib/totp';
 
 type VerifyTwoFactorResponse = {
     id: string;
