+++ src/app/profile/data/setup-two-factor.ts
@@ -4,8 +4,8 @@ import { cache } from 'react';
 import { eq } from 'drizzle-orm';
 import { db } from '@/db';
 import { users } from '@/db/schema/users';
-import { generateSecret, generateOtpauth } from '@/lib/two-factor';
-import { decryptSecret, encryptSecret } from '@/lib/encrypt';
+import { decryptSecret, encryptSecret } from '@/lib/encryption';
+import { generateSecret, generateOtpauth } from '@/lib/totp';
 
 export type SetupTwoFactorResponse = {
     otpauth: string;
